(fix):  move setUser from homepage to navbar
